05 controller refatoracao

+++ src/main/java/br/com/igbr/portfolioApi/controller/GalleryController.java
@@ -1,6 +1,7 @@
 package br.com.igbr.portfolioApi.controller;
 
 import br.com.igbr.portfolioApi.dto.GalleryDTO;
+import br.com.igbr.portfolioApi.dto.TagDTO;
 import br.com.igbr.portfolioApi.model.GalleryModel;
 import br.com.igbr.portfolioApi.repository.GalleryRepository;
 import br.com.igbr.portfolioApi.service.GalleryService;
@@ -8,8 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/gallery")
@@ -17,9 +20,6 @@ import java.util.List;
 public class GalleryController {
     @Autowired
     private GalleryService service;
-
-    @Autowired
-    private GalleryRepository repository;
 
     @GetMapping
     public ResponseEntity<List<GalleryDTO>> getAll(){
@@ -45,7 +45,11 @@ public class GalleryController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public ResponseEntity delete(@PathVariable Long id) {
+        Optional<GalleryDTO> dto = service.findById(id);
+        if(dto.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         service.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 }
